+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0-alpha.3
+
+* Division migrator: Treat most slash operations within argument lists as
+  division.
+* Division migrator: Only migrate slash operations to slash-list calls in a
+  non-plain-CSS context.
+
 ## 1.0.0-alpha.2
 
 * Internal changes only.